Merge test-v3

+++ commons/tester.py
@@ -19,7 +19,7 @@ class Tester:
         self.metric.reset()
         for idx, batch in enumerate(loop):
             data = batch[0].to(self.device)
-            target = batch[1].long().to(self.device)
+            target = batch[1].to(self.device)
             target = torch.argmax(target, 1)
             with torch.no_grad():
                 pred = self.model(data)
